requirements.txt
Django==1.4
M2Crypto==0.21.1
PyBrowserID==0.5.0
certifi==0.0.8
chardet==1.0.1
distribute==0.6.24
psycopg2==2.4.5
requests==0.11.2
wsgiref==0.1.2
